requirements.txt
streamlit
plotly
streamlit_pandas_profiling
# baostock